Universal solution for datetime of last stream date
Previously, the last stream date showed different results when viewed from different timezones. The current solution converts the stream date itself to UTC, which solves that problem. Apparently when working with dates, it is most effective to convert everything to the same timezone.

Fix issue with createVodObject function, where it hangs due to not resolving in the case of not finding the file tag. It now resolves with a 0, which is then filtered out before returning. Unsure how to resolve without returning any item to the array.